requirements.txt
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==40.0.2
Deprecated==1.2.13
distlib==0.3.6
docutils==0.19
gcloud==0.18.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.21.0
idna==3.4
jwcrypto==1.4.2
jws==0.1.3
Kivy==2.1.0
Kivy-Garden==0.1.5
kivymd==1.1.1
multidict==6.0.4
oauth2client==4.1.3
Pillow==9.4.0
pkce==1.0.3
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
Pygments==2.14.0
pyparsing==3.0.9
Pyrebase4==4.6.0
python-jwt==4.0.0
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
urllib3==1.26.14
wrapt==1.15.0
